Config 분리
